transformer/requirements.txt
python-dotenv
pymongo
minio
beautifulsoup4
lxml